requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
joblib==1.4.2
scikit-learn==1.5.1

sqlalchemy==2.0.32
asyncpg==0.29.0

tensorflow==2.16.1
